requirements.txt
alembic==1.4.2
aniso8601==6.0.0
atomicwrites==1.4.0
attrs==20.2.0
awscli==1.18.142
Babel==2.8.0
botocore==1.18.1
cffi==1.14.2
Click==7.0
colorama==0.4.3
cryptography==3.1
docutils==0.15.2
ecdsa==0.14.1
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.1
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
urllib3==1.25.10
Werkzeug==0.15.4
WTForms==2.3.3
zipp==3.1.0
